Refactoring and cleanup.
