--- modules/mamba_trainer.py
@@ -1,26 +1,0 @@
-from transformers import Trainer
-import torch
-import os
-import json
-
-class MambaTrainer(Trainer):
-    def compute_loss(self, model, inputs, return_outputs=False):
-        input_ids = inputs.pop("input_ids")
-        lm_logits = model(input_ids).logits
-
-        labels = input_ids.to(lm_logits.device)
-        shift_logits = lm_logits[:, :-1, :].contiguous()
-        labels = labels[:, 1:].contiguous()
-
-        loss_fct = torch.nn.CrossEntropyLoss()
-        lm_loss = loss_fct(shift_logits.view(-1, shift_logits.size(-1)), labels.view(-1))
-
-        return lm_loss
-
-    def save_model(self, output_dir, _internal_call: bool = False):
-        if not os.path.exists(output_dir):
-            os.makedirs(output_dir)
-            
-        open(f"{output_dir}/config.json", "w").write(self.model.config.to_json_string())
-        torch.save(self.model.state_dict(), f"{output_dir}/pytorch_model.bin")
-        self.tokenizer.save_pretrained(output_dir)
